Separate Apache Shiro from server module
Motivation:
For now, Apache Shiro provides an authentication service for Central Dogma server. It's enough for LDAP, JDBC or Active Directory based authentication, but not useful for some authentication methods like OAuth or SAML. So it would be better to split the current authentication layer into abstraction authentication layer and its implementation layer so that we can replace the authenticaion implementation with the new one easily.

Modifications:
- Remove Apache Shiro from server dependencies, and make a new module.
- Add c.l.c.server.authentication package and provide AuthenticationProvider and AuthenticationProviderFactory interfaces for authentication implementations.
- Add session manager classes which have features provided by FileBasedSessionDAO before.

Result:
We can support OAuth or SAML easily.